requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
requests==2.32.3

# Memoria vectorial
qdrant-client==1.9.1
sentence-transformers==3.0.1
numpy>=1.24.0

# Grafo
networkx==3.3

# LLM tools
duckduckgo-search==6.2.13

# API
python-multipart==0.0.9

# PDF
pdfplumber==0.11.0
pypdf==4.3.1

# Utils
tqdm==4.66.4

pyserial==3.5